+++ libnavui-util/src/main/java/com/mapbox/navigation/ui/utils/internal/Provider.kt
@@ -1,6 +1,12 @@
 package com.mapbox.navigation.ui.utils.internal
 
+/**
+ * An object capable of providing instances of type T.
+ */
 fun interface Provider<T> {
+    /**
+     * Provides an instance of T.
+     */
     fun get(): T
 }